# Replace Load32/Load64 with BinaryPrimitives equivalents