Optimize encoding. Below is the before/after for encoding Mark.Twain-Tom.Sawyer.txt, ~35% improvement.
```
| Method |      Mean |    Error |   StdDev |
|------- |----------:|---------:|---------:|
| Before | 129.27 us | 0.835 us | 0.781 us |
|  After |  84.57 us | 0.357 us | 0.298 us |
```